Rename vertex_ai_demo recipe and vertex_parallel package to gemini_ prefix
Renames the recipe directory (vertex_ai_demo -> gemini_ai_demo) and its
client package (vertex_parallel -> gemini_parallel, distribution name
vertex-parallel-grounding -> gemini-parallel-grounding), updating all
imports, docs, and the lockfile accordingly. Also removes the redundant
Markdown export of the enrichment cookbook, keeping the notebook as the
single source of truth.
