+++ python-recipes/gemini_ai_demo/.gitignore
