+++ python-recipes/gemini_ai_demo/.python-version
