+++ python-recipes/gemini_ai_demo/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "vertex-parallel-grounding"
+name = "gemini-parallel-grounding"
 version = "0.1.0"
 description = "Vertex AI Gemini integration with Parallel web search grounding"
 readme = "README.md"
@@ -47,7 +47,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/vertex_parallel"]
+packages = ["src/gemini_parallel"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
@@ -55,7 +55,7 @@ asyncio_mode = "auto"
 asyncio_default_fixture_loop_scope = "function"
 
 [tool.coverage.run]
-source = ["src/vertex_parallel"]
+source = ["src/gemini_parallel"]
 branch = true
 
 [tool.coverage.report]
