+++ python-recipes/gemini_ai_demo/src/gemini_parallel/client.py
