+++ python-recipes/gemini_ai_demo/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for gemini-parallel integration."""
