--- python-recipes/vertex_ai_demo/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for vertex-parallel integration."""
