+++ website/cookbook.json
@@ -116,8 +116,8 @@
       "featured": false,
       "title": "Vertex AI Grounding",
       "description": "Ground Gemini on Vertex AI with the Parallel Search API for current, cited responses. Marketplace and BYOK auth.",
-      "repoUrl": "https://github.com/parallel-web/parallel-cookbook/tree/main/python-recipes/vertex_ai_demo",
-      "websiteUrl": "https://github.com/parallel-web/parallel-cookbook/tree/main/python-recipes/vertex_ai_demo",
+      "repoUrl": "https://github.com/parallel-web/parallel-cookbook/tree/main/python-recipes/gemini_ai_demo",
+      "websiteUrl": "https://github.com/parallel-web/parallel-cookbook/tree/main/python-recipes/gemini_ai_demo",
       "creators": ["parallel-web"],
       "imageUrl": null,
       "tags": ["search", "vertex-ai", "python"]
